+++ src/math/ax.f90
@@ -42,7 +42,7 @@ module ax_product
   !> Base type for a matrix-vector product providing \f$ Ax \f$
   type, public, abstract :: ax_t
    contains
-     procedure(ax_compute), nopass, deferred :: compute
+     procedure(ax_compute), pass(this), deferred :: compute
      procedure(ax_compute_vector), pass(this), deferred :: compute_vector
   end type ax_t
 
@@ -52,9 +52,12 @@ module ax_product
      !! @param object The matrix-vector product type to be allocated.
      !! @param full_formulation Whether to use the formulation with the full
      !! viscous stress tensor, not assuming constant material properties.
-     module subroutine ax_helm_factory(object, full_formulation)
+     !! @param svv Optional asymmetric SVV data.
+     module subroutine ax_helm_factory(object, full_formulation, svv)
+       use spectral_vanishing_viscosity, only : svv_t
        class(ax_t), allocatable, intent(inout) :: object
        logical, intent(in) :: full_formulation
+       type(svv_t), intent(in), target, optional :: svv
      end subroutine ax_helm_factory
   end interface
 
@@ -68,13 +71,14 @@ module ax_product
   !! @param msh Mesh.
   !! @param Xh Function space \f$ X_h \f$.
   abstract interface
-     subroutine ax_compute(w, u, coef, msh, Xh)
+     subroutine ax_compute(this, w, u, coef, msh, Xh)
        import space_t
        import mesh_t
        import coef_t
        import ax_t
        import rp
        implicit none
+       class(ax_t), intent(in) :: this
        type(space_t), intent(in) :: Xh
        type(mesh_t), intent(in) :: msh
        type(coef_t), intent(in) :: coef