+++ src/math/ax_helm_fctry.f90
@@ -32,7 +32,7 @@
 !
 submodule (ax_product) ax_helm_fctry
   use neko_config, only : NEKO_BCKND_SX, NEKO_BCKND_XSMM, &
-       NEKO_BCKND_DEVICE
+       NEKO_BCKND_DEVICE, NEKO_BCKND_CUDA, NEKO_BCKND_HIP
   use ax_helm_device, only : ax_helm_device_t
   use ax_helm_xsmm, only : ax_helm_xsmm_t
   use ax_helm_sx, only : ax_helm_sx_t
@@ -40,6 +40,11 @@ submodule (ax_product) ax_helm_fctry
   use ax_helm_cpu, only : ax_helm_cpu_t
   use ax_helm_full_cpu, only : ax_helm_full_cpu_t
   use ax_helm_full_device, only : ax_helm_full_device_t
+  use ax_helm_svv_cpu, only : ax_helm_svv_cpu_t
+  use ax_helm_svv_device, only : ax_helm_svv_device_t
+  use ax_helm_svv_full_cpu, only : ax_helm_svv_full_cpu_t
+  use ax_helm_svv_full_device, only : ax_helm_svv_full_device_t
+  use spectral_vanishing_viscosity, only : svv_t
   use utils, only : neko_error
   implicit none
 
@@ -50,9 +55,11 @@ contains
   !! @param object The matrix-vector product type to be allocated.
   !! @param full_formulation Whether to use the formulation with the full
   !! viscous stress tensor, not assuming constant material properties.
-  module subroutine ax_helm_factory(object, full_formulation)
+  !! @param svv Optional asymmetric SVV data.
+  module subroutine ax_helm_factory(object, full_formulation, svv)
     class(ax_t), allocatable, intent(inout) :: object
     logical, intent(in) :: full_formulation
+    type(svv_t), intent(in), target, optional :: svv
 
     if (allocated(object)) then
        deallocate(object)
@@ -63,10 +70,51 @@ contains
          call neko_error("Full stress formulation is only available &
                         &on the CPU and device")
       else if (NEKO_BCKND_DEVICE .eq. 1) then
-         allocate(ax_helm_full_device_t::object)
+         if (present(svv)) then
+            if (NEKO_BCKND_CUDA .ne. 1 .and. NEKO_BCKND_HIP .ne. 1) then
+               call neko_error("Asymmetric full-stress SVV is only " // &
+                    "available on CPU, CUDA, and HIP backends")
+            end if
+            allocate(ax_helm_svv_full_device_t::object)
+            select type (operator => object)
+            type is (ax_helm_svv_full_device_t)
+               operator%svv => svv
+            end select
+         else
+            allocate(ax_helm_full_device_t::object)
+         end if
       else
-         allocate(ax_helm_full_cpu_t::object)
+         if (present(svv)) then
+            allocate(ax_helm_svv_full_cpu_t::object)
+            select type (operator => object)
+            type is (ax_helm_svv_full_cpu_t)
+               operator%svv => svv
+            end select
+         else
+            allocate(ax_helm_full_cpu_t::object)
+         end if
       end if
+    else if (present(svv)) then
+       if (NEKO_BCKND_SX .eq. 1 .or. NEKO_BCKND_XSMM .eq. 1) then
+          call neko_error("Asymmetric SVV is not available with the SX or " // &
+               "XSMM backend")
+       else if (NEKO_BCKND_DEVICE .eq. 1) then
+          if (NEKO_BCKND_CUDA .ne. 1 .and. NEKO_BCKND_HIP .ne. 1) then
+             call neko_error("Asymmetric SVV is only available on CPU, " // &
+                  "CUDA, and HIP backends")
+          end if
+          allocate(ax_helm_svv_device_t::object)
+          select type (operator => object)
+          type is (ax_helm_svv_device_t)
+             operator%svv => svv
+          end select
+       else
+          allocate(ax_helm_svv_cpu_t::object)
+          select type (operator => object)
+          type is (ax_helm_svv_cpu_t)
+             operator%svv => svv
+          end select
+       end if
     else
        if (NEKO_BCKND_SX .eq. 1) then
           allocate(ax_helm_sx_t::object)